flask/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.1.1
werkzeug==2.3.7
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-exporter-otlp==1.20.0
python-dotenv==1.0.0